pyproject.toml
[project]
name = "ptit-csdldpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "librosa>=0.11.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    "pyarrow>=23.0.1",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "streamlit>=1.55.0",
    "python-multipart>=0.0.22",
    "tqdm>=4.67.3",
]
